+++ CHANGELOG.md
@@ -9,6 +9,8 @@ bumped for multiple releases during one month.
 ### [Unreleased]
 ##### Fixed
 - Order Confirmation event will now correctly fire for orders with coupons.
+##### Updated
+- Rename sendEmail function to trackEvent
 
 ### [21.7.0] - 2021-7-27
 ##### Updated